+++ src/packages/babel-preset-kiwicom/src/__tests__/__fixtures__/dynamic-import.js
@@ -1,0 +1,20 @@
+// @flow strict
+
+/* eslint-disable */
+
+const React = require('react');
+
+const BoldTodoRenderer = React.lazy(() =>
+  import(
+    // $FlowExpectedError: cannot resolve module
+    './BoldTodoRenderer'
+  ),
+);
+
+module.exports = function MyComponent() {
+  return (
+    <React.Suspense fallback={<div>Loading...</div>}>
+      <BoldTodoRenderer />
+    </React.Suspense>
+  );
+};