+++ src/packages/babel-preset-kiwicom/src/index.js
@@ -78,10 +78,11 @@ module.exports = (
         'jsx',
         'flow',
         // Enable parsing of (not transpilation) - necessary for Flow target:
-        'classProperties',
-        'nullishCoalescingOperator',
-        'objectRestSpread',
-        'optionalChaining',
+        'classProperties', // https://github.com/tc39/proposal-class-public-fields
+        'dynamicImport', // https://github.com/tc39/proposal-dynamic-import
+        'nullishCoalescingOperator', // https://github.com/babel/proposals/issues/14
+        'objectRestSpread', // https://github.com/tc39/proposal-object-rest-spread
+        'optionalChaining', // https://github.com/tc39/proposal-dynamic-import
 
         // see: https://babeljs.io/docs/en/babel-parser#plugins
         // Candidates: numericSeparator, classPrivateProperties, classPrivateMethods